Button 1.0
Finished the button for the map which now allows for the user to see and click/tap on it to change the map that the user will use. This will be here until I first decide if I want the button on the map or for it to be a setting in the settings menu, and if the latter, then I will remove the button an put it into the settings.